Add 'list' command
The list command lists all activities in the time order they occurred.
The output contains the dates and times the activities covered along with
their names.  The list command takes optional filters that limit the
activities displayed.  For example, filtering the list by a date limits
the resulting activities to those that match the date.  By doing this,
you can easily see all of the activities form a a certain day.  You
can filter by more than one string at a time, so you could show only
activities for +project-x and 2015-05-24.  When filters are provided,
a total number of hours is calculated for the matching activities.
This makes the list command an excellent way to fill out a timecard
or run reports for time spent on a specific project.